+++ .gitattributes
@@ -21,8 +21,8 @@
 /package.json                                    	linguist-generated
 /src/backend/catalog-builder/catalog-builder.ts  	linguist-generated
 /src/backend/discovery/discovery.ts              	linguist-generated
-/src/backend/docgenerator/docgenerator.ts        	linguist-generated
 /src/backend/ingestion/ingestion.ts              	linguist-generated
+/src/backend/transliterator/transliterator.ts    	linguist-generated
 /src/monitored-certificate/certificate-monitor.ts	linguist-generated
 /src/monitoring/http-get-function.ts             	linguist-generated
 /tsconfig.eslint.json                            	linguist-generated
