+++ .gitignore
@@ -54,8 +54,8 @@ yarn-error.log*
 !/src/__tests__
 !/src/backend/catalog-builder/catalog-builder.ts
 !/src/backend/discovery/discovery.ts
-!/src/backend/docgenerator/docgenerator.ts
 !/src/backend/ingestion/ingestion.ts
+!/src/backend/transliterator/transliterator.ts
 !/src/monitored-certificate/certificate-monitor.ts
 !/src/monitoring/http-get-function.ts
 !/tsconfig.eslint.json
