+++ package.json
@@ -29,8 +29,8 @@
     "dev:destroy": "npx projen dev:destroy",
     "bundle:catalog-builder": "npx projen bundle:catalog-builder",
     "bundle:discovery": "npx projen bundle:discovery",
-    "bundle:docgenerator": "npx projen bundle:docgenerator",
     "bundle:ingestion": "npx projen bundle:ingestion",
+    "bundle:transliterator": "npx projen bundle:transliterator",
     "bundle:certificate-monitor": "npx projen bundle:certificate-monitor",
     "bundle:http-get-function": "npx projen bundle:http-get-function",
     "projen": "npx projen",
