+++ src/backend/index.ts
@@ -1,5 +1,5 @@
 export * from './catalog-builder';
 export * from './discovery';
 export * from './ingestion';
-export * from './docgenerator';
+export * from './transliterator';
 export * from './discovery';
